sign up and floats fixed

+++ server_folder/model.py
@@ -1,4 +1,4 @@
-## RUN THE APP
+# RUN THE APP
 from server_folder import db
 
 ## SEED DATABASE
@@ -17,7 +17,7 @@ class User(db.Model):
     password = db.Column(db.String(64), nullable=False)
     first_name = db.Column(db.String(64), nullable=False)
     last_name = db.Column(db.String(64), nullable=False)
-    budget = db.Column(db.Float, nullable=False, default=0.00)
+    budget = db.Column(db.Numeric(10,2), nullable=False, default=0.00)
     bill_color = db.Column(db.String(64), nullable=False, default='#955525')
     new_month_switch = db.Column(db.Boolean, nullable=False, default=False)
 
@@ -28,7 +28,7 @@ class Category(db.Model):
     category_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     color = db.Column(db.String(64), nullable=False)
     title = db.Column(db.String(64), nullable=False)
-    limit = db.Column(db.Float, nullable=False)
+    limit = db.Column(db.Numeric(10,2), nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'))
 
 class Expense(db.Model):
@@ -38,7 +38,7 @@ class Expense(db.Model):
     expense_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     name = db.Column(db.String(255), nullable=False)
     location = db.Column(db.String(255), nullable=False)
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(10,2), nullable=False)
     date = db.Column(db.DateTime, nullable=False)
     category_id = db.Column(db.Integer, db.ForeignKey('categories.category_id'))
     user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'))
@@ -57,7 +57,7 @@ class Bill(db.Model):
 
     bill_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     name = db.Column(db.String(255), nullable=False)
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(10,2), nullable=False)
     payed = db.Column(db.Boolean, nullable=False, default=False)
     user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'))
